fix: Fix the log_dir
If the log_dir is /var/log/kolla, and when we
deploy with devstack, we only chown the permission
for /var/log/kolla/skyline with stack user.
But the permission for /var/log/kolla is still root.
So it can not be access by stack user.
We will both change the log_dir in skyline-console
project.

Change-Id: I852c44cda2f9879aa8d8e405bb39bba13f35b8ca