feat: Add time_expired into cookie
Now the expired time for skyline and keystone is not the same
value. Sometime, when the token of keystone is valid but the
jwt token of skyline-apiserver is invalid.
The skyline-console will send the request to backend service
and then get 401 from skyline.
So we add time_expired into cookie, then skyline-console can
check whether time expired or not.

Change-Id: Id1d3a83eb433c18e88828115e8bd744151fb14f7