Multiple CI fixes
These need to be squashed into a single commit in order to make the CI
pass. The skyline-functional-devstack is still made non-voting since
the fixes in [0] are also needed in order to make it work again after
the switch to Ubuntu 24.04.

Fix some pep8 and typing errors.

Bump aiosqlite version as per upper constraints:

aiosqlite is u-c with 0.20.0
- https://github.com/openstack/requirements/blob/8a8abf5abf6035066f9071f2baec1049a2e250ee/upper-constraints.txt#L156

We need to bump it in requirements.txt also otherwise below requirements
conflict occur:

The conflict is caused by:
    skyline-apiserver 5.1.0.dev2 depends on aiosqlite<=0.17.0
    The user requested (constraint) aiosqlite===0.20.0

[0] https://review.opendev.org/c/openstack/skyline-console/+/939685

Change-Id: I2dababab7c79e7c37983c1702178419c88e014a5