Add config params
* access_log_file
* error_log_file

While generating the Nginx config let's not
use hard coded log directory but instead allow
users to specify either absolute or relative
log file paths.

The following scenarios are possible:

1) log_dir is specified (or not specified and default log_dir is used)
   and *_log_file is NOT specified. In this case the default *_log_file
   is used and it does not break existing deployments.
2) log_dir is speficied (or not ...) and *_log_file is also specified
   as a relative path. In this case the resulting log file path is
   calculated as {log_dir}/{*_log_file}
3) log_dir is specified (or not ...) and *_log_file is specified
   as an absolute path. In this case the log_dir is ignored and the
   resulting file path is calculated as {*_log_file}

The change is important for deployments when users would like Nginx to
write logs to STDOUT. This is typical when services deployed on top
of Kubernetes. In this case they can specify *_log_file=/dev/stdout.

Change-Id: I21f7f4eb4a9aa2bebd9911aedb1f62afa47bd79e