Include tenant_id in user context, too
See https://bugs.launchpad.net/skyline-console/+bug/2115420:
neutron-vpnaas uses tenant_id in the default policy rules but we
do not supply that information in the policy endpoint.

It's not enough to have tenant_id in the target, to make the rules
match we also need it in the user context.

Change-Id: Ibaa6a47f21e6c4f7a5b54ace881ac89e4013da64
Signed-off-by: Uwe Jäger <uwe.jaeger@valiton.com>